refactor: create module for mouse handlers
