+++ .eslintrc.yml
@@ -8,6 +8,7 @@ parserOptions:
   sourceType: module
 rules:
   no-console: 0
+  no-param-reassign: "off"
   import/extensions:
     - error
     - ignorePackages
